Refactor getTokenFrom function into a middleware, add get route for single blog entry -exercise 4.20

+++ controllers/blogs.js
@@ -3,14 +3,6 @@ const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-const getTokenFrom = (req) => {
-  const authorization = req.get('authorization');
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7);
-  }
-  return null;
-};
-
 blogsRouter.get('/', async (req, res, next) => {
   try {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 });
@@ -22,8 +14,7 @@ blogsRouter.get('/', async (req, res, next) => {
 
 blogsRouter.post('/', async (req, res, next) => {
   try {
-    const token = getTokenFrom(req);
-    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
     if (!decodedToken.id) {
       return res.status(401).json({ error: 'token missing or invalid' });
     }
@@ -47,7 +38,19 @@ blogsRouter.post('/', async (req, res, next) => {
   }
 });
 
-blogsRouter.put('/:id', async (req, res, next) => {  
+blogsRouter.get('/:id', async (req, res, next) => {
+  try {
+    const blog = await Blog.findById(req.params.id);
+    if (!blog) {
+      return res.status(404).end();
+    }
+    res.json(blog);
+  } catch (error) {
+    next(error);
+  }
+});
+
+blogsRouter.put('/:id', async (req, res, next) => {
   const blogContent = {
     title: req.body.title,
     author: req.body.author,
@@ -55,15 +58,16 @@ blogsRouter.put('/:id', async (req, res, next) => {
     likes: req.body.likes || '0',
   };
   try {
-    const token = getTokenFrom(req);
-    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
     if (!decodedToken.id) {
       return res.status(401).json({ error: 'token missing or invalid' });
     }
-    const blog = await Blog.findById(req.params.id)
-    if (blog.user.toString() !== decodedToken.id){
-      return res.status(401).json({ error: 'wrong user, only the original creator can edit a blog' })
-    }  
+    const blog = await Blog.findById(req.params.id);
+    if (blog.user.toString() !== decodedToken.id) {
+      return res
+        .status(401)
+        .json({ error: 'wrong user, only the original creator can edit a blog' });
+    }
     const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blogContent, {
       new: true,
       runValidators: true,
@@ -71,24 +75,25 @@ blogsRouter.put('/:id', async (req, res, next) => {
     });
     res.json(updatedBlog);
   } catch (error) {
-    next(error)
+    next(error);
   }
- 
 });
 
 blogsRouter.delete('/:id', async (req, res, next) => {
-  const token = getTokenFrom(req);
-   
   try {
-    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
     if (!decodedToken.id) {
       return res.status(401).json({ error: 'token missing or invalid' });
     }
-    const blog = await Blog.findById(req.params.id)
-
-    if (blog.user.toString() !== decodedToken.id){
-      return res.status(401).json({ error: 'wrong user, only the original creator can delete a blog' })
-    }  
+    const blog = await Blog.findById(req.params.id);
+    if (!blog) {
+      return res.status(204).end();
+    }
+    if (blog.user.toString() !== decodedToken.id) {
+      return res
+        .status(401)
+        .json({ error: 'wrong user, only the original creator can delete a blog' });
+    }
     await Blog.findByIdAndDelete(req.params.id);
     res.status(204).end();
   } catch (error) {
